+++ src/neptune/internal/backends/neptune_backend.py
@@ -18,6 +18,7 @@ __all__ = ["NeptuneBackend"]
 import abc
 from typing import (
     Any,
+    Generator,
     List,
     Optional,
     Tuple,
@@ -58,6 +59,7 @@ from neptune.internal.id_formats import (
 from neptune.internal.operation import Operation
 from neptune.internal.utils.git import GitInfo
 from neptune.internal.websockets.websockets_factory import WebsocketsFactory
+from neptune.typing import ProgressBarType
 
 
 class NeptuneBackend:
@@ -154,6 +156,7 @@ class NeptuneBackend:
         container_type: ContainerType,
         path: List[str],
         destination: Optional[str] = None,
+        progress_bar: Optional[ProgressBarType] = None,
     ):
         pass
 
@@ -164,6 +167,7 @@ class NeptuneBackend:
         container_type: ContainerType,
         path: List[str],
         destination: Optional[str] = None,
+        progress_bar: Optional[ProgressBarType] = None,
     ):
         pass
 
@@ -231,6 +235,7 @@ class NeptuneBackend:
         path: List[str],
         index: int,
         destination: str,
+        progress_bar: Optional[ProgressBarType],
     ):
         pass
 
@@ -303,7 +308,11 @@ class NeptuneBackend:
         types: Optional[List[ContainerType]] = None,
         query: Optional[NQLQuery] = None,
         columns: Optional[List[str]] = None,
-    ) -> List[LeaderboardEntry]:
+        limit: Optional[int] = None,
+        sort_by: str = "sys/creation_time",
+        ascending: bool = False,
+        progress_bar: Optional[ProgressBarType] = None,
+    ) -> Generator[LeaderboardEntry, None, None]:
         pass
 
     @abc.abstractmethod