+++ tests/unit/neptune/new/internal/backends/test_hosted_file_operations.py
@@ -102,19 +102,12 @@ class TestCommonHostedFileOperations(HostedFileOperationsHelper):
         swagger_mock = self._get_swagger_mock()
         exp_uuid = str(uuid.uuid4())
 
-        pre_download_hook = MagicMock()
-        download_iter_hook = MagicMock()
-        post_download_hook = MagicMock()
-
         # when
         download_file_attribute(
             swagger_client=swagger_mock,
             container_id=exp_uuid,
             attribute="some/attribute",
             destination=None,
-            pre_download_hook=pre_download_hook,
-            download_iter_hook=download_iter_hook,
-            post_download_hook=post_download_hook,
         )
 
         # then
@@ -125,7 +118,9 @@ class TestCommonHostedFileOperations(HostedFileOperationsHelper):
             query_params={"experimentId": str(exp_uuid), "attribute": "some/attribute"},
         )
         store_response_mock.assert_called_once_with(
-            download_raw.return_value, None, pre_download_hook, download_iter_hook, post_download_hook
+            download_raw.return_value,
+            None,
+            None,
         )
 
     @patch("neptune.internal.backends.hosted_file_operations._store_response_as_file")
@@ -139,18 +134,11 @@ class TestCommonHostedFileOperations(HostedFileOperationsHelper):
         swagger_mock = self._get_swagger_mock()
         download_id = str(uuid.uuid4())
 
-        pre_download_hook = MagicMock()
-        download_iter_hook = MagicMock()
-        post_download_hook = MagicMock()
-
         # when
         download_file_set_attribute(
             swagger_client=swagger_mock,
             download_id=download_id,
             destination=None,
-            pre_download_hook=pre_download_hook,
-            download_iter_hook=download_iter_hook,
-            post_download_hook=post_download_hook,
         )
 
         # then
@@ -160,7 +148,9 @@ class TestCommonHostedFileOperations(HostedFileOperationsHelper):
             headers={"Accept": "application/zip"},
         )
         store_response_mock.assert_called_once_with(
-            download_raw.return_value, None, pre_download_hook, download_iter_hook, post_download_hook
+            download_raw.return_value,
+            None,
+            None,
         )